support for making digital forensics bags
- refactor audio & digital baggers to separate subclasses rather than
just the metadata extractor

- option parser to make source path optional

- example digital forensics bag with zeroed-out content

- remove non-functional test in chipmunk_bag_spec - was trying to test a
behavior in the base class that failed
