+++ lib/chipmunk_digital_bagger.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "find"
+require "audio_mets"
+require "chipmunk_metadata_error"
+require "chipmunk_bagger"
+
+class ChipmunkDigitalBagger < ChipmunkBagger
+
+  # use default implementation - common tags & expect bag to already exist in place
+
+end
