+++ spec/support/fixtures/digital/pre-chipmunk/bagit.txt
@@ -1,0 +1,2 @@
+BagIt-Version: 0.97
+Tag-File-Character-Encoding: UTF-8
