+++ spec/support/fixtures/digital/pre-chipmunk/manifest-md5.txt
@@ -1,0 +1,11 @@
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/bulk-extractor-output/report.xml
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/bulk-extractor-output/telephone_histogram.txt
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/bulk-extractor-output/telephone.txt
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/bulk-extractor-output/rfc822.txt
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/bulk-extractor-output/hex.txt
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/39015092233538_MediaFront.JPG
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/39015092233538_MediaBack.JPG
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/reporting-output/reports/premis.xml
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/metadata/reporting-output/fiwalk-output.xml
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/image/39015092233538.E01
+d41d8cd98f00b204e9800998ecf8427e data/39015092233538/image/39015092233538.info
