+++ spec/support/fixtures/digital/pre-chipmunk/manifest-sha1.txt
@@ -1,0 +1,11 @@
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/bulk-extractor-output/report.xml
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/bulk-extractor-output/telephone_histogram.txt
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/bulk-extractor-output/telephone.txt
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/bulk-extractor-output/rfc822.txt
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/bulk-extractor-output/hex.txt
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/39015092233538_MediaFront.JPG
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/39015092233538_MediaBack.JPG
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/reporting-output/reports/premis.xml
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/metadata/reporting-output/fiwalk-output.xml
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/image/39015092233538.E01
+da39a3ee5e6b4b0d3255bfef95601890afd80709 data/39015092233538/image/39015092233538.info
